Update index.html
Edit to dynamically change the start url in the manifest to the current URL. This way, the user doesn't have to set up their subscription twice when adding push (specifically to the home screen on iOS), just grant permissions to send the push for their first topic. May be a hacky way to accomplish this, but it's pretty much the only way to achieve a dynamic start url for the app's manifest that is compliant across browsers.